+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/MainActivity.kt
@@ -19,6 +19,7 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(findViewById(R.id.andesui_nav_bar))
         supportActionBar?.title = resources.getString(R.string.andesui_demoapp_app_name)
 
+        setupCheckbox()
         setupRadioButton()
         setupSnackbar()
         setupTags()
@@ -31,6 +32,12 @@ class MainActivity : AppCompatActivity() {
         setupAndesSpecsWeb()
     }
 
+    private fun setupCheckbox() {
+        andesui_checkbox.setOnClickListener {
+            launchIntent("meli://andes/checkbox")
+        }
+    }
+  
     private fun setupRadioButton() {
         andesui_radiobutton.setOnClickListener {
             launchIntent("meli://andes/radiobutton")
@@ -75,8 +82,8 @@ class MainActivity : AppCompatActivity() {
 
     private fun setupWhatsNew() {
         andesui_demoapp_changelog.setupPrimaryAction(
-            getString(R.string.andesui_demoapp_whatsnew_main_action),
-            View.OnClickListener { launchIntent("meli://andes/whats-new") }
+                getString(R.string.andesui_demoapp_whatsnew_main_action),
+                View.OnClickListener { launchIntent("meli://andes/whats-new") }
         )
     }