+++ demoapp/src/main/java/com/mercadolibre/android/andesui/demoapp/feature/MessageShowcaseActivity.kt
@@ -67,9 +67,7 @@ class MessageShowcaseActivity : AppCompatActivity() {
 
         private fun addDynamicMessage(inflater: LayoutInflater): View {
             val layoutMessagesChange = inflater.inflate(
-                    R.layout.andesui_message_showcase_change,
-                    null,
-                    false
+                    R.layout.andesui_message_showcase_change, null, false
             ) as ScrollView
 
             val hierarchySpinner: Spinner = layoutMessagesChange.findViewById(R.id.hierarchy_spinner)
@@ -165,9 +163,7 @@ class MessageShowcaseActivity : AppCompatActivity() {
 
         private fun addStaticMessages(inflater: LayoutInflater): View {
             val layoutMessages = inflater.inflate(
-                    R.layout.andesui_message_showcase,
-                    null,
-                    false
+                    R.layout.andesui_message_showcase, null, false
             ) as ScrollView
 
             val button = layoutMessages.findViewById<AndesButton>(R.id.button)