+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v2.6.0
 ## Added
+- Checkbox Component.
 - RadioButton Component.
 
 # v2.5.1